changed double to long

+++ experimente/experiment_01/warehousemanagement/src/main/java/ch/hslu/wipro/micros/warehousemanagement/repository/FakeWarehouseRepository.java
@@ -5,20 +5,20 @@ import ch.hslu.wipro.micros.common.ArticleDto;
 import java.util.HashMap;
 
 public class FakeWarehouseRepository implements WarehouseRepository {
-    private HashMap<Double, ArticleDto> inventory;
+    private HashMap<Long, ArticleDto> inventory;
 
     public FakeWarehouseRepository() {
         inventory = new HashMap<>();
-        inventory.put(0d, new ArticleDto("Apple", 100d));
-        inventory.put(1d, new ArticleDto("Banana", 150d));
-        inventory.put(2d, new ArticleDto("Pineapple", 150d));
-        inventory.put(3d, new ArticleDto("Hamster", 2050d));
-        inventory.put(4d, new ArticleDto("Cat", 15050d));
-        inventory.put(5d, new ArticleDto("Dog", 35050d));
+        inventory.put(0L, new ArticleDto("Apple", 100));
+        inventory.put(1L, new ArticleDto("Banana", 150));
+        inventory.put(2L, new ArticleDto("Pineapple", 150));
+        inventory.put(3L, new ArticleDto("Hamster", 2050));
+        inventory.put(4L, new ArticleDto("Cat", 15050));
+        inventory.put(5L, new ArticleDto("Dog", 35050));
     }
 
     @Override
-    public ArticleOperation getArticleDtoById(double id) {
+    public ArticleOperation getArticleDtoById(long id) {
         if (inventory.containsKey(id)) {
             return new ArticleOperation(true, inventory.get(id));
         }
